Implement best practices when invoking sync/async code
We previously had a mixture of calling sync over async as well
as calling async over sync (commands over implementations). This
may have been convenient, but it had a number of drawbacks:

- It basically was an implementation of two anti-patterns, if you
  will: http://blogs.msdn.com/b/pfxteam/archive/2012/03/24/10287244.aspx
  and http://blogs.msdn.com/b/pfxteam/archive/2012/04/13/10293638.aspx.
  Read those for more info on why what we were doing isn't a good idea.

- And it also introduced a dependency on the Async package, to
  better handle scenarios that could hang an app (sync waiting on
  an async implemetnation)

Now the choice of invocation style for a command is entirely up to the
command handler implementer, which decides up-front if his command is
async or not, and that forces the required API usage on ICommandBus
which now only allows calling ExecuteAsync on commands that are marked
as async. This makes the choice explicit and makes the consumer aware
of potentially long running commands.

If offloading is necessary to unblock a UI thread for example, the
consumer can still take a dependency on the Async package and just
perform an `asyncManager.Run(() => commandBus.ExecuteAsync(command))`
invocation. This will follow both the ParalellFx team's best practices
as well as VS'.