Add a generic type for easy introspection of handlers
Since previously we had no generic common base interface type
across ICommandHandler<TCommand> and IAsyncCommandHandler<TCommand>,
we had to specifically check for handlers that didn't implement either
generic interface.

Now we can instead check for types that implement the new
IExecutableCommandHandler<TCommand> to easily inspect the TCommand,
regardless of whether the handler is sync or async, and whether it
returns a value or not.