Split interfaces from implementations
This will make it easier to have unchanged contracts/interfaces nuget packages
and potentially evolving implementations. It also makes it super-obvious that
the implementations aren't somehow "blessed" over a custom implementation, and
most of the time consumers within an extensible host app, will only need a
reference to the interfaces, no the implementations.

Also cleaned-up the AssemblyInfo.cs even more to avoid unnecessary files (
AssemblyTitle and AssemblyDescription just used the AssemblyName for the most
part, so make that part of GlobalAssemblyInfo.cs.