Remove externally dependency on System.Reactive
By building the VSIX-bundled Merq.Core as an ILRepack-ed assembly,
we avoid depending on a VS-installed Rx, which may or may not be
compatible with what we use.

Since we only use Rx just for the Subject<T> type and Observable
overloads, as an internal implementation detail of the event
stream, we can safely embed the dependency and internalize it
entirely.