Remove dependency on System.Reactive for EventStream
The dependency on Rx is problematic since other extensions can distribute the same assemblies.
In addition, the dynamic way in which Rx locates the platform assemblies makes this scenario particularly problematic since a previously loaded platform assembly would break this. In addition, our IL merging means that the resulting platform assembly name does not match the one we IL merged, so it's not even possible to locate our version instead.

So instead of trying to solve the complicated problem, we step back and determine exactly why we need (if at all) a dependency on Rx. It turns our that we can trivially implement ourselves the couple classes we actually were using from Rx, with minimal cost and 100% behavior compatibility (all existing unit tests pass with no changes). So we can remove the IL merged project as well as the entire Rx dependency.

Consumers of IEventStream can still depend on Rx to do improved querying, but our implementation no longer needs to rely on it.