Unify the way versions are determined in interface packages
Both Merq and Merq.Async had their own way of setting the version
information. Now that all the other packages use a unified approach,
it made sense to remove the customizations in both and leverage the
single approach. Had to extend the base approach to account for things
those packages did.

Also declared netstandard1.6 as the target, so we can remove package
dependencies that aren't needed (System.Runtime and System.Threading.Tasks).