Don't use floating versions of GitInfo/MSBuilder
The GitInfo one is prone to having behavior differences
across versions, even if code didn't change on our side,
so it's better to let the shared .props specify a fixed
version.

Moved the src/Merq.targets contents to the Directory.Build.targets
so that all projects get it without explicitly importing it.