[cg] Remove legacy netfx-System.StringResources package (#7)
We need to provide license information of all the 3rd party packages we depend on, and this legacy one is missing its license. Since it doesn't seem to be maintained, it's better to remove it. On top of that, it was adding much value.

Fixes https://dev.azure.com/devdiv/DevDiv/_workitems/edit/1552394
AB#1552394