Create a unified MessageBus implementation based on IServiceProvider
This new implementation is more performant than the original CommandBus for public types, since it relies on dynamic dispatch which is faster, as meassured by the new benchmarks project. For the non-public types scenarios, it performs like the original command bus, which is great too and only incurs a downcast.