Improve analyzer and fixer for generic TCommand invocations
The extension methods on IMessageBus allows passing the TCommand directly, as long as they have a parameterless constructor.

In this case, the analyzer should still report the same diagnostics as for a parametrized method call, and the codefix should be the same as before (turn Execute > await ExecuteAsync and viceversa).