Add analzer and codefix for command handler implementations
Implementing command handlers is a bit tricky when the command returns a value, since you need to specify both the command type and its return type in the ICommandHandler<T, TResult> declaration. Changes in the command definition (adding a return type or changing it) result in hard to read compiler errors which can be discouraging.

The new analyzer will detect these situations:
- Missing return type for a non-void command
- Mismatched return type

And we offer code fixers that fix both scenarios, including rewriting the Execute method to have the correct return type.

Closes #37