Add support for hierarchical record creation from generated factories
We now invoke factories automatically whenever they are
found.

A new diagnostic is reported if a `Create(dynamic value)` factory
method is found in the record but it's not accessible to the
generated factory in the current assembly. The diagnostic is
nevertheless a warning since in that case we still generate our
custom factory, but users will not get the benefit of invoking
their custom factory (although this might be by design if it's
just a coincidence in the method name and signature they
chose for an entirely unrelated purpose).

Closes #47