Condition the VSIX-related target to VSSDK install
Currently, the targets are mainly intended to avoid including the main library in VSIX projects that can otherwise just use the built-in Merq in VS.

By adding a condition on the VSSDK-provided `VsSDKInstall` property, we make sure we never even attempt to run the target (or locate its dependent target) in any other scenario.

We also don't really need the target to be constrained to NS2, so we remove that too to simplify the project.