Fix automapper dependency
If we don't specify the target framework, the packaging project will just take a dependency with no framework. This doesn't work when we *also* have framework-specific dependencies, and will cause the dependency to be lost entirely when nuget performs a restore.

In other words, if you have non-framework specific dependencies AND framework specific dependencies, you will only get the latter in the restore graph.