Switch to ValueTask<T> to avoid allocations from commands
When commands complete before a method returning, we can avoid a Task allocation altogether by using ValueTask instead of Task. Since the goal of executing a command is typically to retrieve a value and use that right-away, the optimization makes sense in general.

Aligns with the new NotifyAsync for events too.

Fixes #125